+++ bin/habtracker-daemon.py
@@ -1252,7 +1252,11 @@ def createDirewolfConfig(callsign, l, configdata):
                 f.write("\n\n")
                 f.write("######### beaconing configuration #########\n")
                 f.write("GPSD\n");
-                f.write("TBEACON sendto=" + str(channel) + " delay=0:30 every=" + str(configdata["beaconlimit"]) + "  altitude=1    via=WIDE1-1,WIDE2-1,EOSS      symbol=/k    comment=\"EOSS Tracker - " + str(callsign) + "\"\n");
+                if configdata["includeeoss"] == "true":
+                    eoss = ",EOSS"
+                else:
+                    eoss=""
+                f.write("TBEACON sendto=" + str(channel) + " delay=0:30 every=" + str(configdata["beaconlimit"]) + "  altitude=1    via=WIDE1-1,WIDE2-1" + str(eoss) + "      symbol=" + str(configdata["symbol"]) + "    comment=\"" + str(configdata["comment"]) +  "\"\n");
                 f.write("SMARTBEACONING " + str(configdata["fastspeed"]) + " " + str(configdata["fastrate"]) + "      " + str(configdata["slowspeed"]) + " " + str(configdata["slowrate"]) + "     " + str(configdata["beaconlimit"]) + "     " + str(configdata["fastturn"]) + " " + str(configdata["slowturn"]) + "\n");
                 f.write("###########################################\n\n")
 
@@ -1302,8 +1306,8 @@ def direwolf(configfile="", logfile="", e = None):
     try:
         # We open the logfile first, for writing
         l = open(logfile, "w")
-        
-        # Run the dirwolf command
+
+        # Run the direwolf command
         p = sb.Popen(df_command, stdout=l, stderr=l)
 
         # Wait for it to finish