+++ www/readconfiguration.php
@@ -44,67 +44,10 @@
     else 
 	    $configuration = json_decode($configJSON, true);
 
-
-
-    // Check timezone 
-    if (!isset($configuration["timezone"]))
-	    $configuration["timezone"] = $defaults["timezone"];
-
-    // Check callsign 
-    if (!isset($configuration["callsign"]))
-	    $configuration["callsign"] = $defaults["callsign"];
-
-    // Check lookbackperiod
-    if (!isset($configuration["lookbackperiod"]))
-	    $configuration["lookbackperiod"] = $defaults["lookbackperiod"];
-
-    // Check iconsize
-    if (!isset($configuration["iconsize"]))
-	    $configuration["iconsize"] = $defaults["iconsize"];
-
-    // Check plottracks
-    if (!isset($configuration["plottracks"]))
-	    $configuration["plottracks"] = $defaults["plottracks"];
-
-    // Check SSID
-    if (!isset($configuration["ssid"]))
-	    $configuration["ssid"] = $defaults["ssid"];
-
-    // Check igating
-    if (!isset($configuration["igating"]))
-	    $configuration["igating"] = $defaults["igating"];
-
-    // Check beaconing
-    if (!isset($configuration["beaconing"]))
-	    $configuration["beaconing"] = $defaults["beaconing"];
-
-    // Check fastspeed
-    if (!isset($configuration["fastspeed"]))
-	    $configuration["fastspeed"] = $defaults["fastspeed"];
-
-    // Check slowspeed
-    if (!isset($configuration["slowspeed"]))
-	    $configuration["slowspeed"] = $defaults["slowspeed"];
-
-    // Check fastrate
-    if (!isset($configuration["fastrate"]))
-	    $configuration["fastrate"] = $defaults["fastrate"];
-
-    // Check slowrate
-    if (!isset($configuration["slowrate"]))
-	    $configuration["slowrate"] = $defaults["slowrate"];
-    
-    // Check beaconlimit
-    if (!isset($configuration["beaconlimit"]))
-	    $configuration["beaconlimit"] = $defaults["beaconlimit"];
-
-    // Check fastturn
-    if (!isset($configuration["fastturn"]))
-	    $configuration["fastturn"] = $defaults["fastturn"];
-
-    // Check slowturn
-    if (!isset($configuration["slowturn"]))
-	    $configuration["slowturn"] = $defaults["slowturn"];
+    foreach(array_keys($defaults) as $key) {
+        if(!array_key_exists($key, $configuration))
+            $configuration[$key] = $defaults[$key];
+    }
 
     //header("Content-Type:  application/json;");
     printf ("%s", json_encode($configuration));